finished RecordService, refactored services, corrected GroupRepository

+++ src/main/java/com/example/kma_schedule/database/repository/RecordRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecordRepository extends CrudRepository<Record, Long> {
+public interface RecordRepository extends CrudRepository<Record, Integer> {
 
 }
